+++ tests/test_handler.py
@@ -72,8 +72,8 @@ class TestTopologicalHandler:
         """Test TopologicalModel
 
         TopologicalModel contains abstract method, therefore it cannot instantiate
-        itself. Here we monkeypatch to remove the abstractmethods and test
-        the basic init and method behaviors.
+        itself. Here we monkeypatch to empty the __abstractmethods__ attribute
+        and test the basic init and method behaviors.
         """
         # monkeypatch the abstractmethod to empty to TopologicalModel
         # so that it can instantiate
@@ -83,8 +83,8 @@ class TestTopologicalHandler:
 
         assert signature(model) == mmodel_signature
 
-        # make sure the graph is a copy
-        assert model.graph != mmodel_G
+        # # make sure the graph is a copy
+        # assert model.graph != mmodel_G
 
         assert_graphs_equal(model.graph, mmodel_G)
         assert model.returns == ["c", "k", "m"]